ApiCliente

+++ src/main/java/Fundamentos/de/Analisis/de/Sistemas/repositorios/ClienteRepository.java
@@ -8,4 +8,4 @@ import Fundamentos.de.Analisis.de.Sistemas.modelos.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer>{};
+public interface ClienteRepository extends JpaRepository<Cliente, Long>{};
